+++ src/main/java/com/jxm/yiti/req/WxQuestionSaveReq.java
@@ -5,7 +5,7 @@ public class WxQuestionSaveReq {
 
     private String title;
 
-    private Long categoryId;
+    private Integer categoryId;
 
     private Integer like;
 
@@ -51,11 +51,11 @@ public class WxQuestionSaveReq {
         this.title = title;
     }
 
-    public Long getCategoryId() {
+    public Integer getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(Long categoryId) {
+    public void setCategoryId(Integer categoryId) {
         this.categoryId = categoryId;
     }
